Added support for multitask learning.
The code is now modified from the original seq2seq-attn implementation
to incorporate support for multitask learning. We can now train
the model to jointly train two prediction tasks, for example
trainslation and morphological tag prediction. The only constraint
is that both tasks must have the same target lengths in the same
instance. An additional hyperparameter is added to weigh between
the two tasks.